Refactors table
